+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-## Upcoming
+- Script was confusing fn's `context` with script's `sandbox`.
+- script `sandbox` is now called `scope`.
 
 ## 1.0.0
 - First working version with code grabbed from `requirehit/browser` repo